+++ Validators/FatturaElettronicaHeaderValidator.cs
@@ -37,10 +37,14 @@ namespace FatturaElettronica.Validators
 
         private static bool HeaderValidateAgainstError00476(FatturaElettronicaHeader header)
         {
-            var idCedente = header.CedentePrestatore.DatiAnagrafici.IdFiscaleIVA;
-            var idCommittente = header.CessionarioCommittente.DatiAnagrafici.IdFiscaleIVA;
+            var idCedente = header.CedentePrestatore.DatiAnagrafici.IdFiscaleIVA.IdPaese;
+            var idCommittente =
+                header.CessionarioCommittente.DatiAnagrafici.IdFiscaleIVA is null ||
+                header.CessionarioCommittente.DatiAnagrafici.IdFiscaleIVA.IsEmpty()
+                    ? ""
+                    : header.CessionarioCommittente.DatiAnagrafici.IdFiscaleIVA.IdPaese;
 
-            return (idCedente.IdPaese == "IT" || (idCommittente is not null && idCommittente.IdPaese == "IT"));
+            return idCedente == "IT" || idCommittente is "IT" or "";
         }
     }
 }